refactor(armature): split parser/element into cohesive modules
